+++ CHANGES.rst
@@ -17,6 +17,9 @@ Bug fixes:
 - Fix adapter lookup on Python 3.
   [ale-rt]
 
+- Test using plone.app.contenttypes instead of Archetypes.
+  [davisagli]
+
 4.1.7 (2018-03-12)
 ------------------